requirements.txt
fastapi
sqlalchemy
pymysql
uvicorn
python-dotenv
psycopg2-binary
